requirements.txt
streamlit
geopandas
shapely
spotdl
